fix: bit-equality with mkShell in common cases
`config.shellHook = optionalString false foo` declares `shellHook` to be
an empty string, which is merged with the option's default value of an empty
string, separated by `'\n'`, because the option has the type `lines`.
This results in a final value of `shellHook = "\n"`.

`config = mkIf false { shellHook = foo }` results in a final value of
`shellHook = ""` instead, just as with `mkShell`